refactor: align database type handling with schema updates and improve type conversions
- Replace `u8` and `u32` with `i16` and `i32` in multiple SQL queries to match schema changes.
- Add `From<i16>` implementation for `BackupType` to handle type conversion.
- Update `FromRow` implementations to reflect adjusted database column types.
- Ensure consistent type casting in query bindings for compatibility.

+++ src-actix/authentication/auth_db.rs
@@ -97,7 +97,7 @@ impl UserData {
         T: Into<BitFlags<PermissionFlag>>,
     {
         let users =
-            sqlx::query_as::<_, UserData>(&*sql(r#"SELECT * FROM users WHERE permissions = ?"#)).bind(permissions.into().bits()).fetch_all(pool).await?;
+            sqlx::query_as::<_, UserData>(&*sql(r#"SELECT * FROM users WHERE permissions = ?"#)).bind(permissions.into().bits() as i32).fetch_all(pool).await?;
         Ok(users)
     }
